upd: small fixes
