+++ src/app/app.module.js
@@ -8,6 +8,7 @@ import { StateComponent } from "./core/components/state.component.js";
 import { UserComponent } from "./core/components/user.component.js";
 import { FeaturesComponent } from "./core/components/features.component.js";
 import { AppComponent } from "./app.component.js";
+import { FsReadComponent } from "./features/fs/read.js";
 
 export class AppModule extends Module {
     components = [
@@ -20,6 +21,7 @@ export class AppModule extends Module {
         StateComponent,
         UserComponent,
         AppComponent,
+        FsReadComponent,
     ];
 
     bootstrap = [
