--- src/app/features/featuresList.js
@@ -1,9 +1,0 @@
-import { NavigationUpComponent } from "./navigation/up.component.js";
-import { NavigationLsComponent } from "./navigation/ls.component..js";
-import { NavigationCdComponent } from "./navigation/cd.component.js";
-
-export const featuresList = [
-    NavigationUpComponent,
-    NavigationLsComponent,
-    NavigationCdComponent,
-]
